requirements.txt
django==5.0.2
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
tensorflow==2.15.0
numpy==1.24.3
pandas==2.2.0
scikit-learn==1.3.0
joblib==1.3.2
opencv-python==4.9.0.80 